requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.5
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
idna==3.10
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
